refactored readerBase/tableReader
Separated out readerBase implementation (common to both tableReader and
joinReader) into separate readerbase.go file.
